+++ src/Translations_Updater/Base.php
@@ -25,7 +25,8 @@ if ( ! defined( 'WPINC' ) ) {
  * @package Fragen\Translations_Updater
  * @author  Andy Fragen
  */
-class Base {
+trait Base {
+	use API;
 
 	/**
 	 * Store details of all repositories that are installed.
@@ -69,31 +70,12 @@ class Base {
 	);
 
 	/**
-	 * Variable to hold boolean to load remote meta.
-	 * Checks user privileges and when to load.
+	 * Variable to hold boolean to check user privileges.
 	 *
-	 * @var bool $load_repo_meta
+	 * @var bool
 	 */
-	protected static $load_repo_meta;
+	protected static $can_user_update;
 
-	/**
-	 * Let's get going.
-	 */
-	public function run() {
-		$this->load_hooks();
-	}
-
-	/**
-	 * Load relevant action/filter hooks.
-	 * Use 'init' hook for user capabilities.
-	 */
-	protected function load_hooks() {
-		add_action( 'init', array( &$this, 'init' ) );
-		add_action( 'init', array( &$this, 'background_update' ) );
-
-		add_filter( 'extra_theme_headers', array( &$this, 'add_headers' ) );
-		add_filter( 'extra_plugin_headers', array( &$this, 'add_headers' ) );
-	}
 
 	/**
 	 * Remove hooks after use.
@@ -109,30 +91,15 @@ class Base {
 	 *
 	 * @return bool
 	 */
-	public function init() {
-		global $pagenow;
-
-		$load_multisite       = ( is_network_admin() && current_user_can( 'manage_network' ) );
-		$load_single_site     = ( ! is_multisite() && current_user_can( 'manage_options' ) );
-		self::$load_repo_meta = $load_multisite || $load_single_site;
-
-		// Set $force_meta_update = true on appropriate admin pages.
-		$force_meta_update = false;
-		$admin_pages       = array(
-			'plugins.php',
-			'themes.php',
-			'update-core.php',
-			'update.php',
-		);
-
-		if ( in_array( $pagenow, array_unique( $admin_pages ), true ) ) {
-			$force_meta_update = true;
+	public function load() {
+		if ( ! Singleton::get_instance( 'Init' )->can_update() ) {
+			return false;
 		}
 
-		if ( $force_meta_update ) {
+		if ( self::$can_user_update ) {
 			$this->forced_meta_update_plugins();
 		}
-		if ( $force_meta_update ) {
+		if ( self::$can_user_update ) {
 			$this->forced_meta_update_themes();
 		}
 
@@ -151,7 +118,7 @@ class Base {
 	 * Performs actual plugin metadata fetching.
 	 */
 	public function forced_meta_update_plugins() {
-		if ( self::$load_repo_meta ) {
+		if ( self::$can_user_update ) {
 			Singleton::get_instance( 'Plugin' )->get_remote_plugin_meta();
 		}
 	}
@@ -160,7 +127,7 @@ class Base {
 	 * Performs actual theme metadata fetching.
 	 */
 	public function forced_meta_update_themes() {
-		if ( self::$load_repo_meta ) {
+		if ( self::$can_user_update ) {
 			Singleton::get_instance( 'Theme' )->get_remote_theme_meta();
 		}
 	}